Change everything from Long to double, so you can place shifts not on the grid

+++ optashift-employee-rostering-gwtui/src/main/java/org/optaplanner/openshift/employeerostering/gwtui/client/pages/TimeslotBlob.java
@@ -18,7 +18,7 @@ public abstract class TimeslotBlob implements Blob<LocalDateTime> {
     }
 
     @Override
-    public long getSizeInGridPixels() {
+    public double getSizeInGridPixels() {
         return scale.toGridPixels(getPositionInScaleUnits()
                 .plus(getTimeslot().getDurationOfTimeslot())) - scale.toGridPixels(getPositionInScaleUnits());
     }
